implementing feedback part 2: putting demote_dim_coord_to_aux_coord and promote_dim_coord_to_aux_coord in util.py as functions
