+++ lib/iris/cube.py
@@ -991,162 +991,6 @@ bound=(1994-12-01 00:00:00, 1998-12-01 00:00:00)
         for factory in self.aux_factories:
             factory.update(old_coord, new_coord)
 
-    def promote_aux_coord_to_dim_coord(self, name_or_coord):
-        """
-        Promotes an AuxCoord on the cube to a DimCoord. This AuxCoord must be
-        associated with a single cube dimension. If the AuxCoord is associated
-        with a dimension that already has a DimCoord, that DimCoord gets
-        demoted to an AuxCoord.
-
-        Args:
-
-        * name_or_coord:
-            Either
-
-            (a) An instance of :class:`iris.coords.AuxCoord`
-
-            or
-
-            (b) the :attr:`standard_name`, :attr:`long_name`, or
-            :attr:`var_name` of an instance of an instance of
-            :class:`iris.coords.AuxCoord`.
-
-        For example::
-
-            >>> print cube
-            air_temperature / (K)       (time: 12; latitude: 73; longitude: 96)
-                 Dimension coordinates:
-                      time                    x      -              -
-                      latitude                -      x              -
-                      longitude               -      -              x
-                 Auxiliary coordinates:
-                      year                    x      -              -
-            >>> cube.promote_aux_coord_to_dim_coord('year')
-            >>> print cube
-            air_temperature / (K)       (year: 12; latitude: 73; longitude: 96)
-                 Dimension coordinates:
-                      year                    x      -              -
-                      latitude                -      x              -
-                      longitude               -      -              x
-                 Auxiliary coordinates:
-                      time                    x      -              -
-
-        """
-
-        if isinstance(name_or_coord, basestring):
-            aux_coord = self.coord(name_or_coord)
-        elif isinstance(name_or_coord, iris.coords.Coord):
-            aux_coord = name_or_coord
-        else:
-            # Don't know how to handle this type
-            msg = ("Don't know how to handle coordinate of type {}. "
-                   "Ensure all coordinates are of type basestring or "
-                   "iris.coords.Coord.")
-            msg = msg.format(type(name_or_coord))
-            raise TypeError(msg)
-
-        if aux_coord in self.dim_coords:
-            # nothing to do
-            return
-
-        if aux_coord not in self.aux_coords:
-            msg = ("Attempting to promote an AuxCoord ({}) "
-                   "which does not exist in the cube.")
-            msg = msg.format(aux_coord.name())
-            raise ValueError(msg)
-
-        coord_dim = self.coord_dims(aux_coord)
-
-        if len(coord_dim) != 1:
-            msg = ("Attempting to promote an AuxCoord ({}) "
-                   "which is associated with {} dimensions.")
-            msg = msg.format(aux_coord.name(), len(coord_dim))
-            raise ValueError(msg)
-
-        try:
-            dim_coord = iris.coords.DimCoord.from_coord(aux_coord)
-        except ValueError, valerr:
-            msg = ("Attempt to promote an AuxCoord ({}) fails "
-                   "when attempting to create a DimCoord from the "
-                   "AuxCoord because: {}")
-            msg = msg.format(aux_coord.name(), valerr.message)
-            raise ValueError(msg)
-
-        old_dim_coord = self.coords(dim_coords=True,
-                                    contains_dimension=coord_dim[0])
-
-        if len(old_dim_coord) == 1:
-            self.demote_dim_coord_to_aux_coord(old_dim_coord[0])
-
-        # order matters here: don't want to remove
-        # the aux_coord before have tried to make
-        # dim_coord in case that fails
-        self.remove_coord(aux_coord)
-
-        self.add_dim_coord(dim_coord, coord_dim)
-
-    def demote_dim_coord_to_aux_coord(self, name_or_coord):
-        """
-        Demotes a DimCoord on the cube to an AuxCoord, leaving that
-        dimension anonymous.
-
-        Args:
-
-        * name_or_coord:
-            Either
-
-            (a) An instance of :class:`iris.coords.DimCoord`
-
-            or
-
-            (b) the :attr:`standard_name`, :attr:`long_name`, or
-            :attr:`var_name` of an instance of an instance of
-            :class:`iris.coords.DimCoord`.
-
-        For example::
-
-            >>> print cube
-            air_temperature / (K)       (time: 12; latitude: 73; longitude: 96)
-                 Dimension coordinates:
-                      time                    x      -              -
-                      latitude                -      x              -
-                      longitude               -      -              x
-                 Auxiliary coordinates:
-                      year                    x      -              -
-            >>> cube.demote_dim_coord_to_aux_coord('time')
-            >>> print cube
-            air_temperature / (K)        (-- : 12; latitude: 73; longitude: 96)
-                 Dimension coordinates:
-                      latitude                -      x              -
-                      longitude               -      -              x
-                 Auxiliary coordinates:
-                      time                    x      -              -
-                      year                    x      -              -
-
-        """
-
-        if isinstance(name_or_coord, basestring):
-            dim_coord = self.coord(name_or_coord)
-        elif isinstance(name_or_coord, iris.coords.Coord):
-            dim_coord = name_or_coord
-        else:
-            # Don't know how to handle this type
-            msg = ("Don't know how to handle coordinate of type {}. "
-                   "Ensure all coordinates are of type basestring or "
-                   "iris.coords.Coord.")
-            msg = msg.format(type(name_or_coord))
-            raise TypeError(msg)
-
-        if dim_coord not in self.dim_coords:
-            # nothing to do
-            return
-
-        coord_dim = self.coord_dims(dim_coord)
-
-        self.remove_coord(dim_coord)
-
-        self.add_aux_coord(dim_coord, coord_dim)
-
     def coord_dims(self, coord):
         """
         Returns a tuple of the data dimensions relevant to the given
